backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
openai>=1.0.0
google-cloud-storage==2.14.0
google-auth>=2.40.0
google-auth-oauthlib>=1.0.0
google-genai>=1.16.0
flask-socketio>=5.3.0
websockets>=11.0.0
eventlet>=0.33.0 